(~) Share pin-GUID S-matrix extraction between Verilog-A and PICWave
VerilogAModuleWriter.ExtractSParameters and PicWaveSMatrixEmitter.ExtractAt
had become byte-identical for the pin-GUID lookup path (validate LogicalPins,
look up (in-flow-GUID, out-flow-GUID) in SMatrix.GetNonNullValues(), throw on
drift). Only the policy differed: Verilog-A falls back to a heuristic when
no matrix is registered; PICWave does not.

Extract a single PirSMatrixExtractor.TryExtract that implements the pure
pin-GUID lookup and returns null when no matrix exists — callers apply
their own policy on top.

    VerilogAModuleWriter: 208 -> 186 lines
    PicWaveSMatrixEmitter: 111 ->  61 lines
    PirSMatrixExtractor:   new 82 lines

Net: ~70 lines of duplication removed, error-message strings unified.
Any future export target (Simphony in #494, PICWave-COM, SPICE-Model, …)
plugs into the same shared extractor instead of growing a third copy.

42/42 VerilogA + PICWave tests green locally.
